reorganizes files into folders
